requirements.txt
Django==2.0.2
django-filter==1.1.0
djangorestframework==3.7.7
dlib==19.9.0
future==0.16.0
mysqlclient==1.3.12
numpy==1.14.1
opencv-python==3.4.0.12
Pillow==5.0.0
pytz==2018.3
scipy==1.0.0
